move nestContainedTerms nto its own file
